+++ src/main/java/cc/cassian/raspberry/compat/emi/EmiBeaconBaseRecipe.java
@@ -32,12 +32,12 @@ public class EmiBeaconBaseRecipe implements EmiRecipe {
 
     @Override
     public List<EmiIngredient> getInputs() {
-        return List.of(EmiIngredient.of(Ingredient.of(Items.BEACON)));
+        return BLOCKS;
     }
 
     @Override
     public List<EmiStack> getOutputs() {
-        return BLOCKS;
+        return List.of();
     }
 
     @Override
@@ -52,12 +52,12 @@ public class EmiBeaconBaseRecipe implements EmiRecipe {
 
     @Override
     public void addWidgets(WidgetHolder widgets) {
-        widgets.addGeneratedSlot((r) -> getInputs().get(0), 123, 54, 0);
+        widgets.addSlot(getInputs().get(0), 54, 0);
         int x = 0;
         int y = 25;
         int i = 1;
         for (Map.Entry<String, Integer> s : BEACON_BASE_BLOCKS.entrySet()) {
-            widgets.addGeneratedSlot((r) -> ingredient(s.getKey()), EmiUtil.RANDOM.nextInt(), x, y).recipeContext(this).appendTooltip(Component.translatable("emi.raspberry.beacon_base.amplifier", s.getValue()).withStyle(ChatFormatting.YELLOW));
+            widgets.addSlot(ingredient(s.getKey()), x, y).recipeContext(this).appendTooltip(Component.translatable("emi.raspberry.beacon_base.amplifier", s.getValue()).withStyle(ChatFormatting.YELLOW));
             if (i != 7) {
                 x+=18;
             }
@@ -121,14 +121,14 @@ public class EmiBeaconBaseRecipe implements EmiRecipe {
         return map;
     }
 
-    private static final List<EmiStack> BLOCKS;
+    private static final List<EmiIngredient> BLOCKS;
 
     static {
-        ArrayList<EmiStack> blocks = new ArrayList<>();
+        ArrayList<EmiIngredient> blocks = new ArrayList<>();
         for (String string : BEACON_BASE_BLOCKS.keySet()) {
             Item value = ForgeRegistries.ITEMS.getValue(ResourceLocation.parse(string));
             if (value != null) {
-                blocks.add(EmiStack.of(value.getDefaultInstance()));
+                blocks.add(EmiIngredient.of(Ingredient.of(value.getDefaultInstance())));
             }
         }
         BLOCKS = blocks;