+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Missing translation keys and incorrect models on unfinished content.
 - Corrected Brewin' and Chewin' EMI integration to no longer show incorrect recipes.
+- Beacon EMI integration shows inputs correctly.
 
 ## [1.8.1] - 2025-09-24